pyproject.toml
[tool.poetry]
name = "geographica-longitude"
packages = [
    { include='longitude' }
]
version = "0.9.0"
description = "A longitudinal lib"
authors = ["Geographica <hello@geographica.com>"]
homepage = "https://github.com/GeographicaGS/Longitude"
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
aiohttp = "^3.5"
apispec = "^0.39"
aredis = "^1.1"
asyncio = "^3.4"
carto = "1.4"
cartoframes = "^0.9.0"
environs = "^4.1"
flasgger = "^0.9.2"
flask = "^1.0"
flask_cors = "^3.0"
geolibs-cartoasync = "^0.0.4"
inflect = "^2.1"
marshmallow = "^2.19"
pandas = "^0.24.1"
psycopg2-binary = "2.7.7"
redis = "3.1"
shapely = "^1.6"
sqlalchemy = "1.3.0b2"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
jupyterlab = "^0.35.6"
pytest = "^4.5"
pytest-cov = "^2.7"
pytest-sugar = "^0.9.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
